ref: trade modal
