+++ src/components/TradeModal/TradeModal.tsx
@@ -1,13 +1,14 @@
 import Modal from '@/components/Modal/Modal.tsx';
 import classes from './TradeModal.module.css';
 import BSDInput from '@/components/BSDInput/BSDInput.tsx';
-import { ChangeEvent, memo, useCallback } from 'react';
+import { ChangeEvent, memo } from 'react';
 import BSDButton from '@/components/BSDButton/BSDButton.tsx';
 import { SubmitHandler, useForm } from 'react-hook-form';
 import { tradeSchema, TradeSchema } from '@/schemas/trade.schema.ts';
 import { zodResolver } from '@hookform/resolvers/zod';
-import useStore from '@/stores/useStore.ts';
+// import useStore from '@/stores/useStore.ts';
 import { ModalProps } from '@/types';
+import useHandleTrade from '@/hooks/useHandleTrade.ts';
 
 const defaultValues = {
   cryptoValue: '',
@@ -46,93 +47,108 @@ const TradeModal = ({
     defaultValues,
   });
 
-  const { addTradeHistory, funds, updateFunds } = useStore();
+  // const { addTradeHistory, funds, updateFunds } = useStore();
+  const { handleTrade } = useHandleTrade();
 
   const cryptoValue = watch('cryptoValue');
   const fiatValue = watch('fiatValue');
 
-  const handleTrade = useCallback(
-    (type: 'buy' | 'sell') => {
-      const assetSold = type === 'buy' ? 'EUR' : assetAbbr;
-      const assetBought = type === 'buy' ? assetAbbr : 'EUR';
-      const sellAmount =
-        type === 'buy' ? parseFloat(fiatValue) : parseFloat(cryptoValue);
-      const buyAmount =
-        type === 'buy' ? parseFloat(cryptoValue) : parseFloat(fiatValue);
-
-      if (type === 'buy' && funds.fiat < sellAmount) {
-        setError('root', {
-          message: 'Insufficient funds (€)',
-        });
-        return;
-      }
-
-      const assetFunds = funds[assetAbbr] || 0;
-
-      if (type === 'sell' && sellAmount > assetFunds) {
-        setError('root', {
-          message: `Insufficient funds (${assetAbbr})`,
-        });
-        return;
-      }
-
-      addTradeHistory({
-        type,
-        rate,
-        assetId,
-        assetBought,
-        assetSold,
-        timestamp: new Date().getTime(),
-        sellAmount,
-        buyAmount,
-      });
-
-      if (type === 'buy') {
-        if (!Object.keys(funds).includes(assetAbbr)) {
-          updateFunds({
-            ...funds,
-            fiat: funds.fiat - sellAmount,
-            [assetAbbr]: buyAmount,
-          });
-        } else {
-          updateFunds({
-            ...funds,
-            fiat: funds.fiat - sellAmount,
-            [assetAbbr]: funds[assetAbbr] + buyAmount,
-          });
-        }
-      } else {
-        updateFunds({
-          ...funds,
-          fiat: funds.fiat + buyAmount,
-          [assetAbbr]: funds[assetAbbr] - sellAmount,
-        });
-      }
-
-      reset();
-      onClose();
-    },
-    [
-      cryptoValue,
-      fiatValue,
-      assetId,
-      addTradeHistory,
-      rate,
-      assetAbbr,
-      funds,
-      reset,
-      setError,
-      updateFunds,
-      onClose,
-    ]
-  );
+  // const handleTrade = useCallback(
+  //   (type: 'buy' | 'sell') => {
+  //     const assetSold = type === 'buy' ? 'EUR' : assetAbbr;
+  //     const assetBought = type === 'buy' ? assetAbbr : 'EUR';
+  //     const sellAmount =
+  //       type === 'buy' ? parseFloat(fiatValue) : parseFloat(cryptoValue);
+  //     const buyAmount =
+  //       type === 'buy' ? parseFloat(cryptoValue) : parseFloat(fiatValue);
+  //
+  //     if (type === 'buy' && funds.fiat < sellAmount) {
+  //       setError('root', {
+  //         message: 'Insufficient funds (€)',
+  //       });
+  //       return;
+  //     }
+  //
+  //     const assetFunds = funds[assetAbbr] || 0;
+  //
+  //     if (type === 'sell' && sellAmount > assetFunds) {
+  //       setError('root', {
+  //         message: `Insufficient funds (${assetAbbr})`,
+  //       });
+  //       return;
+  //     }
+  //
+  //     addTradeHistory({
+  //       type,
+  //       rate,
+  //       assetId,
+  //       assetBought,
+  //       assetSold,
+  //       timestamp: new Date().getTime(),
+  //       sellAmount,
+  //       buyAmount,
+  //     });
+  //
+  //     if (type === 'buy') {
+  //       if (!Object.keys(funds).includes(assetAbbr)) {
+  //         updateFunds({
+  //           ...funds,
+  //           fiat: funds.fiat - sellAmount,
+  //           [assetAbbr]: buyAmount,
+  //         });
+  //       } else {
+  //         updateFunds({
+  //           ...funds,
+  //           fiat: funds.fiat - sellAmount,
+  //           [assetAbbr]: funds[assetAbbr] + buyAmount,
+  //         });
+  //       }
+  //     } else {
+  //       updateFunds({
+  //         ...funds,
+  //         fiat: funds.fiat + buyAmount,
+  //         [assetAbbr]: funds[assetAbbr] - sellAmount,
+  //       });
+  //     }
+  //
+  //     reset();
+  //     onClose();
+  //   },
+  //   [
+  //     cryptoValue,
+  //     fiatValue,
+  //     assetId,
+  //     addTradeHistory,
+  //     rate,
+  //     assetAbbr,
+  //     funds,
+  //     reset,
+  //     setError,
+  //     updateFunds,
+  //     onClose,
+  //   ]
+  // );
 
   const submitTrade: SubmitHandler<TradeForm & { type: 'buy' | 'sell' }> = (
     formData,
     event
   ) => {
     event?.preventDefault();
-    handleTrade(formData.type);
+    // handleTrade(formData.type);
+    const res = handleTrade({
+      type: formData.type,
+      fiatValue,
+      cryptoValue,
+      assetId,
+      assetAbbr,
+      rate,
+      setError,
+    });
+
+    if (res) {
+      reset();
+      onClose();
+    }
   };
 
   const isFiatInvalid =
